Update to show an Interstitial.
The annotation @UsedByGodot is what solve the override of the getPluginsMethods.
Old tutorials never comments about this It should be a new feature.
Right now this show An Interstitial.
We probably want to refactor plugin names.
Maybe we could Setup a Demo project to Test the plugin and define A Better Singleton.